+++ lib/csv_localizations.dart
@@ -1,3 +1,4 @@
 library csv_localizations;
 
 export 'src/csv_localizations.dart';
+export 'src/string_ext.dart';