+++ lib/src/string_ext.dart
@@ -1,0 +1,5 @@
+import '../csv_localizations.dart';
+
+extension StringExt on String {
+  String get tr => CsvLocalizations.instance.string(this);
+}